Big advancments.
Implemented sending and receiving simple packets to an address.
Implemented many of the flag options, 'informally'.

Must be restructured to acomodate options that require multithreading and do better code overall.

Did and restructured the parsing to be similar to ping. Flags that require arguments still don't parse them (Next step).