Pinging!
Program restructuration completed.

Now it's time to add the flags.

The way time is handled will need to change. Currently it's made in a blocking way, and very differently for the original ping.

Is not the intention to copy the way ping does it, but the behaviour is different, and it will be necessary to implement flags.